[MNT] Fix _ensure_path and _ensure_int checkers (#19)
* fix typo

* fix tests

* replace error with FileNotFoundError

* fix issues and enable numeric sample_rate

* fix more

* fix typo

* rm unnecessary steps

* cast sounds to float32

* revert to 20.04 image

* fix typo

* fix typo

+++ stimuli/audio/tests/test_sound.py
@@ -204,7 +204,7 @@ def test_io_dtypes(tmp_path):
         data = signal.astype(dtype)
         wavfile.write(fname, 44100, data)
         sound = Sound(fname)
-        assert sound.signal.dtype == np.float64
+        assert sound.signal.dtype == np.float32
         sound.crop(0, 0.1)
         sound.play()
